requirements.txt
pyserial
win23all